add: Added refactor, moving events into seperate folders
