+++ src/events/GuildCreateDeleteEvents/guildCreate.js
@@ -1,0 +1,65 @@
+const { Events, EmbedBuilder } = require('discord.js');
+
+module.exports = {
+    name: Events.GuildCreate,
+    async execute (guild, client) {
+    const logchannelid = client.config.botJoinChannel;
+
+    let theowner = process.env.devid; 
+    let invite;
+    let invalidInv = 'Failed to create an invite for this server!'
+    try {
+        const channel2 = await guild.channels.cache.random()
+        const channelId = channel2.id;
+        invite = await guild.invites.create(channelId);
+    } catch (error) {
+        client.logs.warn(`[INVITE_ERROR] Failed to create an invite for ${guild.name}!`)
+    }
+
+    const inviteLink = invite ? invite.url : invalidInv;
+
+    await guild.fetchOwner().then(({ user }) => { theowner = user; }).catch(() => {});
+
+    let embed = new EmbedBuilder()
+    .setColor('Green')
+    .setTitle(`__**Joined a New Server**__`)
+    .setDescription(`${guild.name} has invited ${client.user.username} into their server`)
+    .addFields(
+        { name: "Guild Info", value: `>>> \`\`\`${guild.name} (${guild.id})\`\`\`` },
+        { name: "Owner Info", value: `>>> \`\`\`${theowner ? `${theowner.tag} (${theowner.id})` : `${theowner} (${guild.ownerId})`}\`\`\`` },
+        { name: "Member Count", value: `>>> \`\`\`${guild.memberCount}\`\`\`` },
+        { name: "Server Number", value: `>>> \`\`\`${client.guilds.cache.size}\`\`\`` },
+        { name: "Server Invite", value: `>>> \`\`\`${inviteLink}\`\`\`` })
+    .setThumbnail(guild.iconURL({ dynamic: true }))
+    .setFooter({ text: `Orbit ${client.guilds.cache.size}`, iconURL: client.user.avatarURL({ dynamic: true }) })
+    .setTimestamp();
+
+    const LogChannel = client.channels.cache.get(logchannelid) || await client.channels.fetch(logchannelid).catch(() => {}) || false;
+    if (LogChannel) LogChannel.send({ embeds: [embed] }).catch(console.warn);
+
+    console.log(`${color.orange}[${getTimestamp()}]${color.reset} [GUILD_CREATE] ${client.user.username} has been added to a new guild. \n${color.orange}> GuildName: ${guild.name} \n> GuildID: ${guild.id} \n> Owner: ${theowner ? `${theowner.tag} (${theowner.id})` : `${theowner} (${guild.ownerId})`} \n> MemberCount: ${guild.memberCount} \n> ServerNumber: ${client.guilds.cache.size} \n> ServerInvite: ${inviteLink}`)
+    }
+}
+
+const color = {
+    red: '\x1b[31m',
+    orange: '\x1b[38;5;202m',
+    yellow: '\x1b[33m',
+    green: '\x1b[32m',
+    blue: '\x1b[34m',
+    pink: '\x1b[38;5;213m',
+    torquise: '\x1b[38;5;45m',
+    purple: '\x1b[38;5;57m',
+    reset: '\x1b[0m'
+}
+
+function getTimestamp() {
+    const date = new Date();
+    const year = date.getFullYear();
+    const month = date.getMonth() + 1;
+    const day = date.getDate();
+    const hours = date.getHours();
+    const minutes = date.getMinutes();
+    const seconds = date.getSeconds();
+    return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
+};
