+++ src/events/GuildCreateDeleteEvents/guildDelete.js
@@ -1,0 +1,53 @@
+const { Events, EmbedBuilder } = require('discord.js');
+
+module.exports = {
+    name: Events.GuildDelete,
+    async execute (guild, client) {
+    const logchannelid = client.config.botLeaveChannel;
+
+    let theowner = process.env.devid;
+
+    await guild.fetchOwner().then(({ user }) => { theowner = user; }).catch(() => {});
+
+    let embed = new EmbedBuilder()
+    .setColor('Red')
+    .setTitle(`__**Left a Server**__`)
+    .setDescription(`${guild.name} has kicked/ban ${client.user.username} out of their server`)
+    .addFields(
+        { name: "Guild Info", value: `>>> \`\`\`${guild.name} (${guild.id})\`\`\`` },
+        { name: "Owner Info", value: `>>> \`\`\`${theowner ? `${theowner.tag} (${theowner.id})` : `${theowner} (${guild.ownerId})`}\`\`\`` },
+        { name: "Member Count", value: `>>> \`\`\`${guild.memberCount}\`\`\`` },
+        { name: "Server Number", value: `>>> \`\`\`${client.guilds.cache.size}\`\`\`` })
+    .setThumbnail(guild.iconURL({ dynamic: true }))
+    .setFooter({ text: `${client.user.username} ${client.guilds.cache.size}`, iconURL: client.user.avatarURL({ dynamic: true }) })
+    .setTimestamp();
+
+    const LogChannel = client.channels.cache.get(logchannelid) || await client.channels.fetch(logchannelid).catch(() => {}) || false;
+    if (LogChannel) LogChannel.send({ embeds: [embed] }).catch(console.warn);
+
+    console.log(`${color.blue}[${getTimestamp()}]${color.reset} [GUILD_DELETE] ${client.user.username} has left a guild. \n${color.blue}> GuildName: ${guild.name} \n> GuildID: ${guild.id} \n> Owner: ${theowner ? `${theowner.tag} (${theowner.id})` : `${theowner} (${guild.ownerId})`} \n> MemberCount: ${guild.memberCount}`)
+    }
+}
+
+const color = {
+    red: '\x1b[31m',
+    orange: '\x1b[38;5;202m',
+    yellow: '\x1b[33m',
+    green: '\x1b[32m',
+    blue: '\x1b[34m',
+    pink: '\x1b[38;5;213m',
+    torquise: '\x1b[38;5;45m',
+    purple: '\x1b[38;5;57m',
+    reset: '\x1b[0m'
+}
+
+function getTimestamp() {
+    const date = new Date();
+    const year = date.getFullYear();
+    const month = date.getMonth() + 1;
+    const day = date.getDate();
+    const hours = date.getHours();
+    const minutes = date.getMinutes();
+    const seconds = date.getSeconds();
+    return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
+};
