+++ src/events/ReadyEvents/ready.js
@@ -1,4 +1,4 @@
-const config = require('../config');
+const config = require('../../config');
 const fs = require('fs');
 const path = require('path');
 const mongoose = require('mongoose');
@@ -46,7 +46,7 @@ module.exports = {
         if (mongoose.connect) {
             client.logs.success('[DATABASE] Connected to MongoDB successfully.');
 
-            const schemaFolder = path.join(__dirname, '../schemas'); 
+            const schemaFolder = path.join(__dirname, '../../schemas'); 
             fs.readdir(schemaFolder, (err, files) => {
                 if (err) {
                     client.logs.error('[ERROR] Error reading schemas folder:', err);
@@ -59,7 +59,7 @@ module.exports = {
         client.logs.info(`[EVENTS] Started loading events...`);
         client.logs.success(`[EVENTS] Loaded ${client.eventNames().length} events.`);
         
-        const triggerFolder = path.join(__dirname, '../triggers'); 
+        const triggerFolder = path.join(__dirname, '../../triggers'); 
         fs.readdir(triggerFolder, (err, files) => {
             if (err) {
                 client.logs.error('Error reading trigger folder:', err);
